+++ hoomd/hpmc/integrate.py
@@ -8,42 +8,6 @@ from hoomd.integrate import _integrator
 import hoomd
 import sys
 
-def _get_sized_entry(base, max_n):
-    """ Get a sized entry.
-
-    HPMC has some classes and functions templated by maximum size. This convenience function returns a class given a
-    base name and a maximum value for n.
-
-    Args:
-        base (string): Base name of the class.
-        max_n (int): Maximum size needed
-
-    Returns:
-        The selected class with a maximum n greater than or equal to *max_n*.
-    """
-
-    # inspect _hpmc.__dict__ for base class name + integer suffix
-    sizes=[]
-    for cls in _hpmc.__dict__:
-        if cls.startswith(base):
-            # append only suffixes that convert to ints
-            try:
-                sizes.append(int(cls.split(base)[1]))
-            except:
-                pass
-    sizes = sorted(sizes)
-
-    if max_n > sizes[-1]:
-        raise ValueError("Maximum value must be less than or equal to {0}".format(sizes[-1]));
-
-    # Find the smallest size that fits size
-    for size in sizes:
-        if max_n <= size:
-            selected_size = size;
-            break;
-
-    return _hpmc.__dict__["{0}{1}".format(base, size)];
-
 class interaction_matrix:
     R""" Define pairwise interaction matrix
 
@@ -368,12 +332,7 @@ class mode_hpmc(_integrator):
         raise NotImplementedError("You are using a shape type that is not implemented! If you want it, please modify the hoomd.hpmc.integrate.mode_hpmc.get_type_shapes function")
 
     def initialize_shape_params(self):
-        shape_param_type = None;
-        # have to have a few extra checks becuase the sized class don't actually exist yet.
-        if isinstance(self, sphere_union):
-            shape_param_type = data.sphere_union_params.get_sized_class(self.capacity);
-        else:
-            shape_param_type = data.__dict__[self.__class__.__name__ + "_params"]; # using the naming convention for convenience.
+        shape_param_type = data.__dict__[self.__class__.__name__ + "_params"]; # using the naming convention for convenience.
 
         # setup the coefficient options
         ntypes = hoomd.context.current.system_definition.getParticleData().getNTypes();
@@ -1091,6 +1050,13 @@ class simple_polygon(mode_hpmc):
 class polyhedron(mode_hpmc):
     R""" HPMC integration for general polyhedra (3D).
 
+    This shape uses an internal OBB tree for fast collision queries.
+    Depending on the number of constituent spheres in the tree, different values of the number of
+    spheres per leaf node may yield different optimal performance.
+    The capacity of leaf nodes is configurable.
+
+    Only triangle meshes and spheres are supported. The mesh must be free of self-intersections.
+
     Args:
         seed (int): Random number seed.
         d (float): Maximum move displacement, Scalar to set for all types, or a dict containing {type:size} to set by type.
@@ -1106,8 +1072,8 @@ class polyhedron(mode_hpmc):
     * *vertices* (**required**) - vertices of the polyhedron as is a list of (x,y,z) tuples of numbers (distance units)
 
         * The origin **MUST** strictly be contained in the generally nonconvex volume defined by the vertices and faces
-        * The origin centered circle that encloses all verticies should be of minimal size for optimal performance (e.g.
-          don't put the origin right next to a face).
+        * The (0,0,0) centered sphere that encloses all verticies should be of minimal size for optimal performance (e.g.
+          don't translate the shape such that (0,0,0) right next to a face).
 
     * *faces* (**required**) - a list of vertex indices for every face
     * *sweep_radius* (**default: 0.0**) - rounding radius applied to polyhedron
@@ -1116,6 +1082,12 @@ class polyhedron(mode_hpmc):
 
         * .. deprecated:: 2.1
              Replaced by :py:class:`interaction_matrix`.
+    * *capacity* (**default: 4**) - set to the maximum number of particles per leaf node for better performance
+        * .. versionadded:: 2.2
+    * *origin* (**default: (0,0,0)**) - a point strictly inside the shape, needed for correctness of overlap checks
+        * .. versionadded:: 2.2
+    * *hull_only* (**default: True **) - if True, only consider intersections between hull polygons
+        * .. versionadded:: 2.2
 
     Warning:
         HPMC does not check that all requirements are met. Undefined behavior will result if they are
@@ -1139,7 +1111,7 @@ class polyhedron(mode_hpmc):
         mc.shape_param.set('A', vertices=[(-0.5, -0.5, -0.5), (-0.5, -0.5, 0.5), (-0.5, 0.5, -0.5), (-0.5, 0.5, 0.5), \
             (0.5, -0.5, -0.5), (0.5, -0.5, 0.5), (0.5, 0.5, -0.5), (0.5, 0.5, 0.5)], faces = faces);
         mc.shape_param.set('B', vertices=[(-0.05, -0.05, -0.05), (-0.05, -0.05, 0.05), (-0.05, 0.05, -0.05), (-0.05, 0.05, 0.05), \
-            (0.05, -0.05, -0.05), (0.05, -0.05, 0.05), (0.05, 0.05, -0.05), (0.05, 0.05, 0.05)], faces = faces);
+            (0.05, -0.05, -0.05), (0.05, -0.05, 0.05), (0.05, 0.05, -0.05), (0.05, 0.05, 0.05)], faces = faces, origin = (0,0,0));
     """
     def __init__(self, seed, d=0.1, a=0.1, move_ratio=0.5, nselect=4, implicit=False, restore_state=False):
         hoomd.util.print_status_line();
@@ -1735,8 +1707,9 @@ class sphere_union(mode_hpmc):
     R""" HPMC integration for unions of spheres (3D).
 
     This shape uses an internal OBB tree for fast collision queries.
-    Depending on the number of constituent spheres in the tree, different values of the number of spheres per leaf
-    node may yield different optimal performance. The capacity of leaf nodes is configurable.
+    Depending on the number of constituent spheres in the tree, different values of the number of
+    spheres per leaf node may yield different optimal performance.
+    The capacity of leaf nodes is configurable.
 
     Args:
         seed (int): Random number seed.
@@ -1745,7 +1718,8 @@ class sphere_union(mode_hpmc):
         move_ratio (float): Ratio of translation moves to rotation moves.
         nselect (int): The number of trial moves to perform in each cell.
         implicit (bool): Flag to enable implicit depletants.
-        max_members (int): Set the maximum number of members in the sphere union. (deprecated in version 2.2)
+        max_members (int): Set the maximum number of members in the sphere union
+            * .. deprecated:: 2.2
         capacity (int): Set to the number of constituent spheres per leaf node. (added in version 2.2)
         restore_state(bool): Restore internal state from initialization file when True. See :py:class:`mode_hpmc`
                              for a description of what state data restored. (added in version 2.2)
@@ -1763,6 +1737,8 @@ class sphere_union(mode_hpmc):
 
         * .. deprecated:: 2.1
              Replaced by :py:class:`interaction_matrix`.
+    * *capacity* (**default: 4**) - set to the maximum number of particles per leaf node for better performance
+        * .. versionadded:: 2.2
 
     Example::
 
@@ -1780,7 +1756,7 @@ class sphere_union(mode_hpmc):
         mc.shape_param.set('B', diameters=[0.05], centers=[(0.0, 0.0, 0.0)]);
     """
 
-    def __init__(self, seed, d=0.1, a=0.1, move_ratio=0.5, nselect=4, implicit=False, max_members=None, capacity=4, restore_state=False):
+    def __init__(self, seed, d=0.1, a=0.1, move_ratio=0.5, nselect=4, implicit=False, max_members=None, restore_state=False):
         hoomd.util.print_status_line();
 
         if max_members is not None:
@@ -1792,16 +1768,16 @@ class sphere_union(mode_hpmc):
         # initialize the reflected c++ class
         if not hoomd.context.exec_conf.isCUDAEnabled():
             if(implicit):
-                self.cpp_integrator = _get_sized_entry('IntegratorHPMCMonoImplicitSphereUnion', capacity)(hoomd.context.current.system_definition, seed)
+                self.cpp_integrator = _hpmc.IntegratorHPMCMonoImplicitSphereUnion(hoomd.context.current.system_definition, seed)
             else:
-                self.cpp_integrator = _get_sized_entry('IntegratorHPMCMonoSphereUnion', capacity)(hoomd.context.current.system_definition, seed)
+                self.cpp_integrator = _hpmc.IntegratorHPMCMonoSphereUnion(hoomd.context.current.system_definition, seed)
         else:
             cl_c = _hoomd.CellListGPU(hoomd.context.current.system_definition);
             hoomd.context.current.system.addCompute(cl_c, "auto_cl2")
             if not implicit:
-                self.cpp_integrator = _get_sized_entry('IntegratorHPMCMonoGPUSphereUnion', capacity)(hoomd.context.current.system_definition, cl_c, seed)
+                self.cpp_integrator = _hpmc.IntegratorHPMCMonoGPUSphereUnion(hoomd.context.current.system_definition, cl_c, seed)
             else:
-                self.cpp_integrator = _get_sized_entry('IntegratorHPMCMonoImplicitGPUSphereUnion', capacity)(hoomd.context.current.system_definition, cl_c, seed)
+                self.cpp_integrator = _hpmc.IntegratorHPMCMonoImplicitGPUSphereUnion(hoomd.context.current.system_definition, cl_c, seed)
 
         # set default parameters
         setD(self.cpp_integrator,d);
@@ -1810,11 +1786,7 @@ class sphere_union(mode_hpmc):
         self.cpp_integrator.setNSelect(nselect);
 
         hoomd.context.current.system.setIntegrator(self.cpp_integrator);
-        self.capacity = capacity
         self.initialize_shape_params();
-
-        # meta data
-        self.metadata_fields = ['capacity']
 
         if implicit:
             self.implicit_required_params=['nR', 'depletant_type']