+++ ChangeLog.md
@@ -27,6 +27,7 @@ Not yet released
 
 *Bug fixes*
 
+* hpmc.integrate.sphere_union() and hpmc.integrate.polyhedron() missed overlaps
 * `hpmc.integrate.sphere_union()` and `hpmc.integrate.polyhedron()` missed overlaps
 * fix alignment error when running implicit depletants on GPU with ntrial > 0